+++ tests/TestData/TestClassA.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Rammewerk\Component\Container\Tests\TestData;
+
+class TestClassA {
+
+    public function __construct() {
+    }
+
+}