+++ tests/TestData/TestClassB.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Rammewerk\Component\Container\Tests\TestData;
+
+readonly class TestClassB {
+
+    public function __construct(public TestClassA $classA) {
+
+    }
+
+}