+++ tests/TestData/TestClassC.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Rammewerk\Component\Container\Tests\TestData;
+
+class TestClassC {
+
+    public function __construct(public string $value) {
+    }
+
+}