+++ tests/TestData/TestClassD.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Rammewerk\Component\Container\Tests\TestData;
+
+class TestClassD {
+
+    public function __construct(
+        private TestClassEInterface $classE
+    ) {
+    }
+
+    public function getE(): bool {
+        return $this->classE->get();
+    }
+
+}