+++ tests/TestData/TestClassE.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Rammewerk\Component\Container\Tests\TestData;
+
+class TestClassE implements TestClassEInterface {
+
+    public function get(): bool {
+        return true;
+    }
+}