+++ tests/TestData/TestClassEInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Rammewerk\Component\Container\Tests\TestData;
+
+interface TestClassEInterface {
+
+    public function get(): bool;
+
+}