+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+0.2.0
+---
+
+- A small improvement in performant when registering shared instances and bindings.
+- Added PHP Unit test to make sure everything container behave as expected
+
 0.1.0
 ---
 Beta release